# Remove isolate processing from IGC parser (Phase 3 - Issue #39)