## Summary
- Remove unnecessary isolate overhead from IGC file parsing
- Simplify polyline generation in flight track widget
- Reduce codebase complexity by ~200 lines

## Context
This PR implements part of Phase 3 optimizations from Issue #39 - removing premature optimizations that add complexity without measurable benefit.

Analysis showed that typical IGC files (<100KB) parse in 5-50ms, making the 17ms isolate overhead counter-productive. The isolate processing was adding complexity and actually slowing down the most common use case.

## Changes

### IGC Parser Simplification
- Removed `parseFileInIsolate()` method and Flutter foundation imports
- Updated all 6 call sites to use direct `parseFile()` method
- Removed static helper method used for isolate execution

### Flight Track Widget Optimization  
- Removed 75+ lines of isolate-based polyline generation code
- Simplified cache key generation (removed crypto dependency)
- Direct polyline generation is now faster for typical tracks

### Code Quality
- Added performance test to verify parsing speed
- Maintained all existing progress indicators for large files
- Preserved all error handling and fallback logic

## Performance Impact

**Before (with isolate):**
- Small file (<100KB): ~22-67ms (5-50ms parsing + 17ms isolate overhead)
- Isolate spawn overhead constant at ~17ms

**After (direct parsing):**
- Small file (<100KB): ~5-50ms (no overhead)
- 17ms faster for typical use case
- Simpler stack traces for debugging

## Testing
- ✅ App runs successfully on emulator
- ✅ IGC import functionality tested and working
- ✅ Flight track visualization displays correctly
- ✅ Progress indicators show during import
- ✅ No UI freezing observed with test files

## Related Issues
- Closes part of #39 (Remove Premature Optimizations)
- Related to #37 (Database Architecture Simplification)

🤖 Generated with [Claude Code](https://claude.ai/code)